Allows adding content via file upload in instancewebapp UI
git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@49 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ s3storageprovider/src/main/java/org/duraspace/s3storage/S3StorageProvider.java
@@ -299,7 +299,7 @@ public class S3StorageProvider implements StorageProvider {
     public void addContent(String spaceId,
                            String contentId,
                            String contentMimeType,
-                           int contentSize,
+                           long contentSize,
                            InputStream content)
     throws StorageException {
         S3Object contentItem = new S3Object(contentId);
